refactor(Repeater): Split RecyclePool/RecyclingElementFactory per winui-port
Splits the recycling pool/factory types per the winui-port file layout. The legacy
RecyclePoolFactory.cs is folded into RecyclePool.Properties.cs. Adds internal
OriginTemplateProperty accessor required by ItemTemplateWrapper. Bodies are 1:1 with
WinUI commit 4b206bce3.

Tracks issue #22908.
